Restructure Python SDK v3 (Part 3)
