+++ deepgram/__init__.py
@@ -3,15 +3,15 @@ __version__ = '0.0.0'
 
 # entry point for the deepgram python sdk
 from .deepgram_client import DeepgramClient
-from .types.deepgram_client_options import DeepgramClientOptions
+from .deepgram_client_options import DeepgramClientOptions
 from .errors import DeepgramError, DeepgramApiError, DeepgramUnknownApiError, DeepgramUnknownError
 
 # live
-from .types.transcription_options import LiveOptions
-from .enums import LiveTranscriptionEvents
+from .clients.live.transcription_options import LiveOptions
+from .clients.live.enums import LiveTranscriptionEvents
 
 # prerecorded
-from .types.transcription_options import PrerecordedOptions
+from .clients.prerecorded.transcription_options import PrerecordedOptions
 
 # manage
-from .clients.manage_client import ManageClient
+from .clients.manage.manage_client import ManageClient
