+++ deepgram/clients/live/helpers.py
@@ -1,31 +1,4 @@
-import json
 from urllib.parse import urlparse, urlunparse, parse_qs, urlencode
-from .types.prerecorded_source import PrerecordedSource
-
-def strip_trailing_slash(url):
-    parsed_url = urlparse(url)
-    path = parsed_url.path.rstrip('/')
-    modified_url = parsed_url._replace(path=path)
-    return urlunparse(modified_url)
-
-url = "https://example.com/path/"
-stripped_url = strip_trailing_slash(url)
-
-def is_buffer_source(provided_source: PrerecordedSource) -> bool:
-    return "buffer" in provided_source
-
-def is_readstream_source(provided_source: PrerecordedSource) -> bool:
-    return "stream" in provided_source
-
-def is_url_source(provided_source: PrerecordedSource) ->  bool:
-    return "url" in provided_source
-
-def is_json(myjson):
-    try:
-        json_object = json.loads(myjson)
-    except ValueError as e:
-        return False
-    return True
 
 def append_query_params(url, params=""):
     parsed_url = urlparse(url)
